Add a script to ensure maintainers
As described in the README, the process we're using here has one manual
step that is often missed: when a new package is uploaded, someone needs
to go and add `FreckleEngineering` as a maintainer on it. Until that
happens, this repository won't manage it.

This script aims to address that by automating the following:

For each package maintained by `HACKAGE_UPLOAD_API_KEY` (currently me)
but _not_ maintained by `HACKAGE_API_KEY` (`FreckleEngineering`), check
to see if:

1. A repository named `freckle/{package}` exists, or
2. A repository with a file named `{package}.cabal` exists

If either of these are true, we can be confident it's a Freckle package
that should have had `FreckleEngineering` added as a maintainer, and we
can just add them right here and now (unless `DRY_RUN`).