Add schedule to CI workflow
This will remove the last manual step when releasing a new
OSS package to Hackage. Now, the maintainers will be
updated automatically within a week.

I chose to do it weekly because it really doesn't matter if it
doesn't happen for a while -- auto-release from CI will 
always work from the jump, and this is more about ensuring
all our engineers have maintainer rights in case manual
intervention is needed for some reason.